start final proyect

+++ Aprendizaje/BuenasPracticas/BuenasPracticas.java
@@ -7,8 +7,8 @@ public class BuenasPracticas {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Ingrese la altura de la pirámide: ");
-        int altura = scanner.nextByte(); 
-        Piramide piramide = new Piramide((byte) altura);
+        int altura = scanner.nextInt(); 
+        Piramide piramide = new Piramide(altura);
         piramide.imprimirPiramide();
         scanner.close(); // Cerrar el scanner al finalizar su uso
     }
